Issue #SB-14499 fix: OOM Issue fix
